requirements.txt
Flask >= 0.11
gunicorn >= 19
PyYAML >= 3.12
Flask-Scss >= 0.5
Flask-Misaka
python-dotenv
cryptography
beautifulsoup4
flask-cache
splinter
